misc corner endpoints
